[feaLib] Compile lookups after assigning lookup indices
Before this change, feaLib had assigned a lookup index at the same
time as compiling each lookup. For chains, the implicit assumption was
that the chain's targets would always come before the contexual chain.
Normally this was indeed the case, but feaLib (and also makeotf)
sometimes merge several chain targets into one single lookup,
and then this assumption was not true anymore.

In the new version, the lookups get compiled in a separate pass,
after assigning lookup indices.

https://github.com/behdad/fonttools/issues/463