[fontTools.__init__] use 'lastResort' handler instead of NullHandler
I think it's better to follow the current Python 3 behavior in the situation
where the user doesn't provide any explicit logging configuration.

Before Python 3.2, when no handlers were found when a logging event occurred,
either the events would be silently dropped (if logging.raiseExceptions is
set to False), or it would throw a single "No handlers could be found" warning
(this was the default behavior).

With Python 3.2 and later, the default behavior is that the events are output
using a "handler of last resort". The latter works like StreamHandler
that writes messages (without formatting) to the current value of sys.stderr
(thus respecting any redirections which may be in effect).