[robofab] Don't round new points
It doesn't seem to be a good idea to round some points in a glyph but
not others. With this change, control points (which don't change)
that were equivalent to other non-curve points before conversion will
remain so. This can be significant when drawing with pen objects.

The rounding has been there since the initial commit; it's not clear
that it serves a purpose.