Revert ".travis.yml: also test with built-in Python 2.7 on OSX 10.11"
This reverts commit a63e1ed5d37797ee361b0a23d09bf8a03aaee23d.

The default OSX Travis worker is now osx10.11-xcode7.3 so we don't need to separately test the latter as it's now default.

Also, since the OSX are getting quite slow lately, we can't afford to test two different python2.7 on OSX; so from now on we just test the system Python 2.7.11 that comes with El Capitan (so no more 2.7.5 on Mavericks).