[varLib] Build avar; build fvar with spec-compliant value ranges
This is a work in progress which hasn't been tested. Also, it assumes
that the interpolation masters have correct usWidthClass and usWeightClass
in their OS/2 table, which is not the case for the current Noto Sans fonts.
Part of the problem is in glyphsLib, and part of it seems in the font
sources; both can be fixed.

https://github.com/googlei18n/fontmake/issues/167