[requirements.txt] add unicodedata2, for 3.5 and below
Note that we only use requirements.txt for running the test suite inside tox.
I didn't add (yet?) unicodedata2 module to 'install_requires' in setup.py.

Python 3.6 does not need it, as it's built-in unicodedata module is the
same as unicodedata2 backport.

PyPy is also excluded because currently there is no pre-compiled wheel
available for that implementation.