[setup.cfg] always use --pyargs option with pytest to re-enable doctests discovery
The --pyargs option makes pytest try interpreting arguments as importable python package names, or if that fails treat them as normal filesystem paths.

Now running `pytest` command will automatically run all the tests in "Tests" folder, as well as the embedded doctests in the installed "fontTools" package.

This is in reply to a comment by Behdad in
https://github.com/fonttools/fonttools/commit/fdbf066a8c41c904cdc290ead89ede765dff0582